Fixed the /slap command
The old velocity vector was selected randomly on all three axes. This
caused a problem if the magnitude of the vector was greater than 9.0,
the maximum velocity, the player would not be thrown.

The new version uses spherical coordinates, selects a random inclination
and azimuth with a fixed velocity of 9.0. This guarantees the player is
slapped with maximum prejudice.